feat: serve dynamic data
