+++ config/database.json
@@ -1,1 +1,44 @@
-[]
+[{"id":
+    "974de548-f223-4c8a-806b-8ab9a20ae1cf",
+    "name":
+    "3d Gear Cube",
+    "description":
+    "Gear meshing cube is certainly unique in operation as compared to the rest.  The cube is also available at an affordable cost and is made from nontoxic, durable plastic.",
+    "imageUrl":
+    "http://wonderfulengineering.com/wp-content/uploads/2016/03/10-Best-Rubiks-cubes-4.jpg",
+    "difficulty":5},
+    {"id":
+        "36b9738b-cb2d-4297-b966-29ea18d151f3",
+        "name":
+        "3x3 Gold Mirror",
+        "description": 
+        "A clever enhancement of the original design. Instead of picking the right colour for each side, you solve the cube based on length and width of each square. This cube has more combinations than the regular one, hence more challenging.", 
+        "imageUrl":
+        "https://wonderfulengineering.com/wp-content/uploads/2016/03/10-Best-Rubiks-cubes-6-610x610.jpg",
+        "difficulty":8},
+    {"id":
+        "243d4f1a-83fc-44d4-9ec6-e87fb4f2203a",
+        "name":
+        "Fluctuation Angle",
+        "description":
+        "One of the most mind-boggling puzzles on our list. While in its cube form the problem might look easy but once you try to solve, it gets complicated quickly. The product though functions smoothly and is made from PVC plastic.",
+        "imageUrl":
+        "http://wonderfulengineering.com/wp-content/uploads/2016/03/10-Best-Rubiks-cubes-9.jpg",
+        "difficulty":7},
+    {"id":
+        "381e80fe-cf8b-4192-b378-1a95f7a776b5",
+        "name":
+        "Toys Skewb Xtreme", 
+        "description":
+        "Though not as difficult as it might look the cube is one of the most popular in the market. Additionally, the manufacturer has taken great care in sustaining the build quality of their product, and it shows as well. The device has durable stickers and the best Rubik’s cube variant you can buy.", 
+        "imageUrl":
+        "http://wonderfulengineering.com/wp-content/uploads/2016/03/10-Best-Rubiks-cubes-3.jpg",
+        "difficulty":8},
+    {"id":
+        "71d2867c-d880-4acf-9598-a5ad76697619",
+        "name":
+        "Skewb Speed",
+        "description":
+        "Much similar to our no.9 but has more swivel edges. The cube has a durable build, and the stickers are attached with glue that adheres properly."
+        ,"imageUrl":
+        "http://wonderfulengineering.com/wp-content/uploads/2016/03/10-Best-Rubiks-cubes-2.jpg","difficulty":3}]
