+++ controllers/create-cube.js
@@ -1,7 +1,5 @@
 const Cube = require('../models/cube');
 
-const newCube = new Cube('Default', 'This is a default cube', 'http://google.com', 1);
+// const newCube1 = new Cube('3d Gear Cube', 'Gear meshing cube is certainly unique in operation as compared to the rest.  The cube is also available at an affordable cost and is made from nontoxic, durable plastic.', 'http://wonderfulengineering.com/wp-content/uploads/2016/03/10-Best-Rubiks-cubes-4.jpg', 5);
 
-console.log(newCube);
-
-newCube.save();
+// newCube5.save();
