+++ controllers/cubes.js
@@ -1,0 +1,14 @@
+const fs = require('fs');
+const path = require('path');
+
+
+const { getCubes } = require('../controllers/database'); 
+const getAllCubes = (callback)=>{
+        getCubes((cubesData)=>{
+         callback(cubesData);
+        });
+}
+
+module.exports = {
+        getAllCubes
+}
