+++ views/details.hbs
@@ -1,12 +1,10 @@
 
     <main>
-      <h1>Gan356 Air SM</h1>
-      <img class="cube" src="https://ae01.alicdn.com/kf/HTB1CSddXRxRMKJjy0Fdq6yifFXa6/Gan-356-Air-SM-3x3-Black-Magic-cube-GAN-Air-SM-Magnetic-3x3x3-Speed-cube-gans.jpg">
+      <h1>{{name}}</h1>
+      <img class="cube" src="{{imageUrl}}">
       <div class="details">
-        <p><span>Description:</span> Magnets in AirSM will not drop, and their positions will be more precise with the
-          Magnets-Snap-On design. With the use of 3mm*2mm magnets, the handfeel will be more stable and more
-          comfortable. P.S. This design is brand new for the AirSM.</p>
-        <p><span>Difficulty level:</span> 3</p>
+        <p><span>Description:</span>{{description}}</p>
+        <p><span>Difficulty level:</span>{{difficulty}}</p>
         <a class="btn" href="/">Back</a>
       </div>
     </main>
