+++ views/index.hbs
@@ -8,6 +8,11 @@
         <input type="number" name="to" class="difficulty" placeholder="Difficulty to...">
         <input type="submit" value="search">
       </form>
+
+      {{#each cubes}}
+        {{> cube}}
+      {{/each}}
+      
       <div class="cube">
         <img class="cube" src="https://ae01.alicdn.com/kf/HTB1CSddXRxRMKJjy0Fdq6yifFXa6/Gan-356-Air-SM-3x3-Black-Magic-cube-GAN-Air-SM-Magnetic-3x3x3-Speed-cube-gans.jpg">
         <p class="name">Gan356 Air SM</p>
